requirements.txt
sentence-transformers
spacy
torch
numpy
scikit-learn
transformers
sentencepiece
pandas
google-cloud-bigquery
db-dtypes
google-cloud-storage
google-cloud-bigquery-storage
requests
tqdm
psutil
vaderSentiment
